Move JuegoSerpientesYEscaleras to clases.py
